myLLM/requirements.txt
accelerate==1.10.1
av==15.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
decord==0.6.0
filelock==3.19.1
fsspec==2025.7.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==11.3.0
psutil==7.0.0
PyYAML==6.0.2
qwen-vl-utils==0.0.8
regex==2025.8.29
requests==2.32.5
safetensors==0.6.2
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.0
typing_extensions==4.15.0
urllib3==2.5.0
